Move transaction annotation types to their own file
This allows importing the types from places that may otherwise have
circular dependencies to the enrichment service. These types likely
should be lifted further to the base background type set, but that can
be done later.
